Fixed bug with rewritten ret code
When I added the lookup code for ret instructions (instead of only
masking), I chose the wrong test instruction.  It should have been
"test byte ptr [eax], 0x3", but it was instead "test al, 0x3".  I
think I did this because of a wrong comment I had placed on the
lookup code for indirect jmp/call instructions.  The wrong
instruction checked whether the address of the target had its low
two bits set, when it should have been checking whether the 4 bytes
at the target had its low two bits set.  It seems that the fact
that I had all my rewritten addresses aligned meant that the target
address basically never had the low bits set, so it never performed
a lookup even when it should have.  This allowed my aligned
code_caller hack to work before this fix, but crash after the fix,
because the instruction after the aligned call had a lower bit set.
Now, I have inserted an extra nop after the call for each driver
test that is using code_caller.  However, driver9 relies on the
miniverse_spatcher.py script to align targets and insert nops.
I quite simply added "nop" after indirect calls in the script, and
driver9 worked.  The reason I found out about this bug in the first
place was that I realized that I had forgotten to insert the nops
after indirect calls in the script, and wondered why my tests had
not crashed.

Now that things seem to be working well with control flow between
the rewritten code and original code, I need to return to dealing
with stale rewritten code, as I believe that the current problems
I'm encountering with Javascript are due to control flowing back
to stale rewritten code.  I think I can just patch all possible
jump targets in the old rewritten code (every 16-byte aligned
instruction) with their updated counterparts in the new code before
discarding the old mapping.  One thing that still worries me is
when main binary (non-rewritten) code tries to return to rewritten
code.  Because the segfault handler will raise an alarm if an
address is not recognized, I should be able to detect it if I set
the stale rewritten regions to be non-executable, but I might need
to keep a patched version of the old mappings around (or read from
the patched rewritten code), with a special case for returning to
stale rewritten code.