Added patching stale rewritten code, changed unknown addr behavior
When a region is rewritten multiple times, any pointers that
continue to point to the old rewritten region may be used to try
to jump back to the stale rewritten code.  For one example of this,
just consider jit code that calls a (non-jit) function that
modifies the caller code (or the code near it).  The non-jit
function will try to return to the stale return address, which may
return into incorrect old code, or return to unallocated memory or
non-executable memory, depending on what we do with the old
rewritten code.  Therefore, we need to handle any calls into the
old code.

My solution here is to patch a target entry for the
new rewritten code at every chunk boundary in the old rewritten
code, and then make the old rewritten code non-executable.  This
alone won't work, because the segfault handler originally would
just abort on unknown segfault addresses, and rewritten code is
not supposed to segfault in normal operation.  Rewritten code also
doesn't have a mapping like the original (non-rewritten) code does,
as original code maps to rewritten code.  Therefore, I change the
behavior of the segfault handler to check if there is a valid entry
at the segfaulted address (indicated by the lowest 2 bits being
set).  The result is masked off like any other lookup, maintaining
the security policy, but also allowing us to redirect attempts to
jump to stale rewritten code.