Fixed alignment bug for call instructions identified as targets
I previously added a nop instruction directly before any
instruction identified as a jump target.  Since the padding
generated for those instructions prior to that point would ensure
they would be aligned to the start of a chunk, this was safe to
do, EXCEPT for call instructions, which were padded to the END
of a chunk.  Adding one extra nop ensured that a call that was
also identified as a jump target would be pushed one byte over the
chunk boundary, and a later return would place control in the
middle of that call instruction after masking off the return
address!

The call to check_target for call instructions had this comment:

TODO: Will this incorrectly align a call identified as jmp target?

This comment was quite prescient, because it did indeed incorrectly
align calls identified as jmp targets.  Perhaps I should check back
through my TODO comments to find out whether I have previously
predicted other bugs before forgetting to verify whether there were
any.