Fixed backup restoration, now align 6-byte ind call instructions
Previously, the way I restored a region backup didn't consider
that the size of a region that had its permissions changed might
not be the same size as the old region and its backup, meaning that
if the backup were larger than the new region it would attempt to
copy more bytes to the region than were writable, and if the
backup were smaller than the new region it would attempt to read
past the end of the backup.  This has been fixed, and it uses the
smaller of the two values to determine how much to copy.  There
likely are edge cases here, since a new smaller region with the
same start address of the previous larger region might take over
the region entry and leave the old part past the end of the new
region "undocumented".  I haven't implemented proper region
splitting, so I know for a fact that it won't work in the general
case for now anyway, but it's something to keep in mind.

I also encountered another example of an indirect jump instruction
in non-rewritten code that calls rewritten code.  This requires
the call instruction to be aligned to the end of a 16-byte "chunk"
because the rewritten code will mask off the return address.  Since
I am not rewriting the original code, every call instruction must
be carefully aligned.  This new example was a new size of
instruction that I had not pattern-matched yet (a 6-byte call with
a 4-byte displacement to add to a register).  The specific
instruction I encountered calculated an address to read from as an
offset from the Global Offset Table.  I made a fairly general
pattern match to recognize instructions of this type (although
it may be possible for instructions to assemble to the form with a
smaller 1-byte displacement, which is something to be cautious
about and be wary of in the future).