Fixed rewriting and aligning of less common call instructions
While most call instructions I have encountered are not unusual,
I found some call instructions that make use of both the mod/rm
byte and a sib byte, along with 1 and 4 byte displacements.  The
instructions may often be garbage instructions, but I noticed a
violation of an assertion on some rewritten code and decided to
make the code for rewriting call instructions more correct.  I also
added a new test to ensure that various sorts of call instructions
are handled, and that they can also be targets themselves.