Added test and broken implementation for overlapping targets
There have been a few cases of overlapping targets that would
interfere with each other and cause a lookup entry to be corrupted.
I at first considered a solution to this that would involve marking
an overlapped target as a non-target to avoid it being looked up
at all, and then using the segfault handler to look up the target
after the rewritten code attempted to jump to the original non-
executable address.  However, this does not actually work, as the
target is always masked off, even if a lookup is not performed.
Eventually, we likely want to provide some complicated special
handling for overlapping pointers, but for now we need a simpler
solution just to ensure everything else works.  Dr. Hamlen has
suggested a temporary solution that should at least allow us to
not worry about overlapping pointers for now.

I also added code to sort relocations for the rewritten code, and
this should be useful even for a more permanent solution, because
it lets us know which targets overlap.  The code that will need to
change will be how we actually handle overlapping targets.