Converted code to temporarily omit masking target addresses
For easier debugging and simpler handling of overlapping targets
in the short term, I have changed the mask for targets to
0xffffffff, which will not change addresses.  This means that a
security policy is not being enforced, but it still can be used to
evaluate our approach without having to worry about overlapping
targets.  For this, jump targets are now tagged with 0x9b (fwait).