Added partial implementation to push old addresses on stack
I have encountered many cases where a return address on the stack
pointing to jit code was being used for calculations.  This is
unfortunate, because for speed I was just using normal call and
return instructions in the rewritten code.  This meant that those
call instructions pushed NEW addresses onto the stack.  If
application code expects return addresses to be certain values or
within certain ranges, then the code can break.  This introspection
is a problem.

This can be solved by converting call instructions into essentially
a push+jmp pair and return instructions into a pop+<lookup>+jmp.
I already had implemented returns performing lookups because of
the possibility of old pointers being on the stack, so this was a
somewhat straightforward change, with the exception of ret <imm>
instructions.  I haven't finished the implementation for that yet.
I want to also write a test for that specifically, and a test
that expects an address to be a specific value.

Defining PUSH_OLD_ADDRESSES enables this setting, and other than
ret <imm> instructions, it appears to work well and passes the
current tests I have written.