Added handling of ret <imm> instructions when pushing old addresses
The last part that I needed to implement to handle converting all
call instructions to push + jmp pairs was handling of ret <imm>,
because I had not yet implemented the code to patch the generated
instructions with the immediate from the return instruction.  This
was needed because the new instructions change the stack pointer
instead of the original ret <imm>.  I am aware of an edge case that
the current code will fail: if the immediate is >= 65531, then the
value will overflow when I add 4 to it to compensate for the extra
4 bytes that a ret instruction naturally moves the stack pointer.
However, I am hoping that such a large immediate would be extremely
rare.  From some testing, I found that switching to pushing old
addresses has made a huge improvement to the number of test cases
that succeed, so the results are very promising!

I have also implemented a test for ret <imm> instructions.