Added fixed offset to target lookups, hackish lookup mem allocation
Many tests were failing because data was embedded in the generated
jit code, meaning that the program actually depended on the jit
code being unmodified.  I was patching jump targets in-place, which
sometimes clobbered pointers that the code expected.  I am
attempting to address this by writing the lookup entries for jump
targets in memory regions a fixed offset away from the jit code
(and a fixed offset away from the original program code, too, as
jit code trying to do a lookup on original code also will apply
the fixed offset).  However, this means we need to have a magical
fixed offset that works for all jit code AND original code and
doesn't collide with any existing regions.  Fortunately, memory
allocation for the jit code seems to be fairly sane and is NOT
all over the memory space for what I have been testing.  However,
one of my own tests, test1-multiple-regions, actually fails because
the fixed offset does not work with the addresses that some of the
generated code is mapped to.  The regions are allocated too high,
and the address wraps around to an unallocated lower address.  This
might have coincidentally worked anyway, but the base address of
my initial jit region was at a low address, and the hackish way I
make sure there is memory at a fixed offset is to allocate a large
region at the fixed offset from the first jit region we encounter.
This assumption fails on test1-multiple-regions, and the size of
the region may also not be large enough for tests that allocate a
lot of memory for jit code.

However, regardless of some of the limitations of this
implementation, I already have significant improvements on the
number of tests that pass.