Fixed bug with code generation for indirect jmp/call, added test
The size of the template assembly code for indirect jmp/call
instructions was hard-coded, and my previous changes changed the
length of the templates.  This did not always result in a crash
due to lucky alignment and the specific instructions that were not
copied over correctly.  I have fixed this and added a test that
tries running some simple code at different alignments to ensure
that it will detect if generated code fails at some alignments.  I
should probably enhance this test later with all types of control
flow instructions so that it properly tests all the possible
variations of code I generate.