Reproduced tricky bug with generated jmp instructions
I haven't fixed the bug yet, but it took quite a lot of effort to
end up reproducing the bug, and I eventually tweaked the relocation
sorting code to make it easier.  Here is the comment I put in the
new test I wrote to reproduce the bug:

The bug I am attempting to reproduce is INCREDIBLY difficult to
reproduce without changing any of the original code, because it
depends on the relocation sorting algorithm to sort several
RELOC_IND relocations with the same target but different offsets
(which results in patching a lookup entry multiple times, and the
last relocation "wins" by having its value as the one that stays.
This should be changed to just filter out multiple RELOC_IND
relocations and just use the lowest one, which should probably be
the original.  All of these relocations SHOULD be safe to use, but
the later relocation offsets point to a direct jump to the original
instruction, so it's a pointless extra step caused by the duplicate
jmp instructions generated by superset disassembly to join together
overlapping sequences with a jump to the original code and avoid
code duplication.  I probably should flag "special" jmp instructions
to avoid generating RELOC_IND relocations for those instructions,
but currently my disassembly code does not pass that info to
instruction generation code).  I ended up changing the relocation
sorting algorithm to save the headache required to get the relocs
in just the right order, but this does mean that this test will work
without actually fixing the root problem by simply making small
tweaks and probably wouldn't trip up an unrelated implementation.
I think that testing having a jump target reached by multiple code
offsets is useful, but it's hard to make a test that covers all the
possible edge cases.

I believe that the root cause of the bug comes from my code not
correctly increasing the values of last_safe_offset and
last_safe_reloc for jmp instructions.  This normally did not cause
an issue because a number of factors have to be true for it to fail:
-The jmp must be the last good instruction in a sequence
-The jmp must be a jmp target or cross a chunk boundary so padding
 is generated
-The generated code for the jmp must be overwritten by other
 generated code due to last_safe_offset not being increased enough
This meant that the bug only would come up in fairly complex code
and made it especially hard to track down.  There may be an easier
way to reproduce this, but it is surprisingly complicated and I had
the most luck with trying to reproduce the bug similiarly to the
real-world example I encountered that led to the bug in the first
place.