Removed region splitting, region backups, patching rewritten code
I discovered an incredibly unusual bug in which a call [edi]
instruction spanned across a page boundary, and the second page it
fell through to had been modified AFTER the call instruction was
originally generated.  This caused the second page to be rewritten
containing only the last byte of the call instruction.  Normally
I identify instructions after a call as a target by default, since
a return instruction will probably eventually return there.
However, not even superset disassembly saves us here, because only
the last byte of the call instruction remains, so the instruction
was not correctly identified as a jump target.

This resulted in me re-evaluating the way that I was dealing with
changes to already rewritten regions.  I was splitting regions in
two if only the latter half of a region was being modified (which
is determined by mprotect calls only changing permissions on some
pages in a region).  I don't think I have yet encountered a page
range in the MIDDLE of a region being changed (as opposed to the
whole thing or a range at the end), so I was simply shrinking the
original region and splitting the modifications to a new region.
However, this obviously fails in the case of the error above.

I originally did this because we only want to rewrite subregions
that have been freshly changed, but the previous implementation
failed in this edge case.  Now I am rewriting the WHOLE region
whenever any part is rewritten.  This is MUCH slower, but more
correct. This is a prime candidate for optimizations later, but it
will introduce complexity that I don't need to deal with while I
am still in the debugging phase.

I also noticed that my changes to no longer modify the original
code regions (since the lookup entries are now at a fixed offset
from the code instead of overwriting the original code) mean that
backing up the original code wasn't needed.  If I want to support
a rewriter mode that still patches the old code, I would need to
keep code for this, but for now I'm going to only support the
version I am working on.

Finally, I also noticed that patching rewritten code (and keeping
old versions of rewritten code in memory) is no longer needed
either, now that I push old return addresses instead of keeping
call instructions in the rewritten code.  Now there should be no
pointer references to the rewritten code potentially floating
around in memory, so I can just free it instead of patching old
rewritten code with lookup entries just in case there is a stale
pointer somewhere.

These changes simplify the management of rewritten regions, which
was getting a bit disorganized, and fix a bug with cross-boundary
instructions, but also result in significantly impaired
performance.  I have also written a test that should fail if these
changes were reverted, but I have not yet confirmed its failure
on the old version.  However, it does work correctly for the
current version.