Added extra buffer space before start of fixed offset buffer
The fixed offset buffer for lookups is allocated the first time a
new code region is added.  If a new region is ever allocated at
a lower address than the first region, the addresses at the fixed
offset from that region could be outside the allocated buffer.  I
have added some extra allocated memory before the start of the
fixed offset to handle some cases of this, although I may need to
expand the size depending on memory allocation behaviors of a code
generator.

I tested this with a Lua benchmark that had failed for this reason,
and it now works.  The benchmark was for some reason not included
in the luajit benchmarks listed on the luajit website, but was
present in its repository.  Because of this, I didn't use it in
performance evaluations, but I wanted to make sure it worked
anyway.

I still need to test if this makes any additional javascript tests
work.