Dynamic alloc of code regions table, spinlocks on wrappers/handler
One test case ended up allocating a lot of code regions, enough
that it wrote past the end of the fixed one page of space I
allocated for the code regions table.  I therefore have
finally implemented dynamic allocation of more pages as needed.

The dynamic re-allocation of more pages moves the address of
the code regions table, as I mmap a new region for it and munmap
the old region.  This led to threading issues because multiple
threads would be in __wrap_mmap/__wrap_mprotect/sigsegv_handler,
and after one thread would obtain a reference to an entry in the
table, another thread would re-allocate the table, leaving that
reference in the other thread pointing to often unmapped memory
(and if not, pointing to garbage data).  Therefore, I have now
implemented spinlocks for these functions that prevent multiple
threads from entering any of these functions, only allowing one
thread at a time and leaving any thread waiting for access in a
spinlock.

Hopefully this might also fix other tests failing
nondeterministically, as I am hoping that by restricting access
to miniverse to one thread at a time will fix those tests.  I am
unsure of the performance hit that the spinlocks might cause, but
that can be optimized later.