Changed signature of is_target to take code base address and size
When determining whether the code at an address is a jump target
or not, I was previously using several heuristics that first
checked whether the address was enough bytes above the start of a
page before attempting to read the bytes prior to that address.
This ensured I wasn't attempting to read from unmapped memory, but
it also eliminated any targets shortly after a page boundary from
being identified by the heuristic.  Changing is_target to receive
the base address of the code region being rewritten (along with its
size, which is currently unused but I included for completeness and
its potential usefulness in some analyses) allows me to safely read
bytes prior to any given instruction without worrying about
accidentally attempting to read unmapped memory or skipping certain
instructions due to not knowing whether the memory before them is
mapped.