Set mapping after gen padding, do not set mapping for special jmps
I encountered a bug that appeared to have a discrepancy between the
mapping entries and the target entries at the fixed offset from the
generated code.  Since it was difficult to determine at what point
this difference would occur, I decided to add a warning in the
rewriter whenever the mapping and a relocation entry for writing
to the fixed offset region differed.

This resulted in a lot of differences.  I found many were due to
the mapping entry being recorded before padding was added for the
instruction, while the relocation was generated after padding was
added.  To make the two agree, I delayed recording the mapping
entry until after the padding was added.

However, this resulted in mapping entries being written multiple
times by subsequent special jmp instructions, generated with a
previously disassembled address so that they would end up being
rewritten to jump to the rewritten address for that original
address.  This broke when control flowed to a jump target from
several disassembly offsets.

I now don't write a mapping entry OR fixed-offset relocation entry
for special jmp instructions.  It appears that one of these were
the cause of the bug, as the particular test case now works.  In
fact, more test cases work than before, although some new test
cases appear to be failing now, and some are still
nondeterministically failing.  Still, this does show improvement.