Fixed bugs with mirroring certain code segments at a fixed offset
The rewritten code reads at a fixed offset from a target address,
because we place a lookup table there to translate old addresses to
rewritten addresses.  For addresses in code that is not rewritten,
we also must allocate memory there, and we tag locations that we
have identified as potential indirect jump targets so that we don't
perform a lookup for those non-rewritten targets.

I originally mirrored all original code segments, which I thought
would allow any code in the program to call the generated code and
then be safely returned to (via the indirect jump of a ret
instruction).  However, Firefox's jit compiler is located within
a shared object (libxul.so), which is loaded into the address space
later.  This meant that an attempt to return from the jit code to
the caller within the shared object caused a crash.

I have changed the mirroring behavior to instead mirror the code
segment that first calls any jit code, and thereby first triggers
the segfault handler.  This allows for any function within that
code segment to call the jit code, which appears to work with how
the jit compiler is designed.

I also encountered a bug with my code segment mirroring for high
addresses, because I was using the function strtol, when I should
have used strtoul.  This broke on code loaded at high addresses.
I fixed this and added an additional test that tests both generated
and caller code loaded at high addresses.