Fixed chunk padding bug
While reviewing the code before making some major changes, I
noticed that some of the generated code for indirect jumps and
calls seemed too long to fit in a chunk.  It seems as if I might
not have noticed that the instructions got too long when I added
the fixed offset for lookups.  This resulted in some instructions
crossing the boundary between chunks, which wasn't noticed because
of my prior changes to disable masking of target addresses.

Besides fixing the offending code, I also added a warning to
gen_padding to signal if I attempt to create padding for a byte
sequence longer than a chunk in the future.