Changed fixed offset lookup entries to be 4 bytes
Now, instead of simply adding a fixed offset to access a lookup
entry, I also multiply the address by 4, expanding the lookup
table to be able to fit a full target address for every address
in the original code region.  This avoids the need to fall back
to the segfault handler for clustered targets.

Preliminary testing shows that some benchmarks may roughly double
their performance, and the number of times the segfault handler
is entered is lower.  However, there seem to be some new crashes
caused by this change, and I'm not sure of the cause yet.