More stat tracking and debug output, disable reloc sorting
This commit contains a lot of changes aimed at trying to debug
issues.  There are several miscellaneous changes, including
increasing the amount of memory allocated for the fixed offset
lookup tables for future rewritten code, which allowed for a
complete successful run of the octane benchmark.

Disabling sorting of relocations, which is not needed now that I
expanded the lookup tables to 4 bytes, has improved performance
considerably.  I think that the large number of relocations
(especially caused by changing my_is_target to return true for any
address) caused relocation sorting to be the biggest cost during
rewriting.

I also now print brackets whenever rewriting starts and finishes.
This showed that each time code was rewritten it usually didn't
take very long, and much of the time is spent in the rewritten
code.  Rewriting may still benefit from some optimizations, but
for now I would say performance has become acceptable.  The problem
at hand now is that I am encountering a lot of uncertainty in
Firefox's js engine at startup.  It seems that if we can get past
initialization, a test is far more likely to succeed, but there is
some crazy nondeterminism that can't be eliminated by any of the
following:
-Disabling ASLR
-Forcing the js engine to run on 1 core
-Disabling helper threads in the js engine
So I'm having trouble tracking down what could be causing these
issues.  There have been multiple crashes that are mostly
repeatable, but rarer cases occasionally appear that I may not
easily be able to reproduce.  One unusual but common crash appears
to be caused by the contents of the VDSO to be clobbered, maybe by
a mmap that I perform, but it has been surprisingly hard to catch
the program in the act.  I have created some fatal error messages
to terminate the program if some of the large mmaps I perform fail,
but it's more concerning if they succeed and clobber existing
memory allocations.  If I can determine that this is happening,
I may need to dynamically determine a fixed offset that will work
for the memory mappings at the start of the program.  This would
not provide a guarantee that there will not be issues with
memory allocations later, but may provide slightly more reliable
startups.