Changed fixed offset to be calculated early at runtime
The issues with a completely fixed offset is that it tends to
eventually interfere with the random allocation of memory mapped
regions.  Now I wait until I have the address of the first segfault
and the address of the caller that attempts to jump to rewritten
code, and attempt to calculate a fixed offset that works with the
memory allocations that exist at that time.  This still has the
potential of not finding an offset without conflicts, but now it
should give an error if it can't.

The giant buffer that is allocated for the fixed-offset lookups
may eventually not be large enough if the memory allocated for the
jit code spreads out far beyond the initial jit code memory
allocation.  I make the assumption that future allocations will
tend to go after the initial allocation, and that they will all
be roughly in the same area.  When this assumption does not hold,
then the fixed offset solution can't work; new allocations at
arbitrary addresses will eventually have allocated memory at the
fixed offset I chose.  However, Firefox's jit engine seems to
generally obey this assumption.  Further testing, of course, is
always needed.

I also fixed a bug caused by the increased size of some of the
rewritten instructions.  It seems that it was possible to generate
more chunks of code than expected and run off the edge of memory
allocated for the new code.  I allocated more memory as soon as
<= 3 chunks were left in the allocated region, but now I think
that it is possible to generate 4 chunks with one instruction.
I have changed the code to allocate more memory as soon as <= 4
chunks are left.