Change code trimming to preserve call instructions
Previously, I considered that since a call instruction would
likely return, it was safe to trim the call if it was followed by
illegal instructions.  However, I finally encountered an example
of code that did just that, which led to some trimmed sequences
being overwritten by the following rewritten instructions.  When
those sequences were executed, it led to an eventual crash.  Now,
I am more cautious and preserve those calls in case they simply
never return.  I also added to test 9 to check for a case similar
to the real-world one I found.