Changed handling of indirect jmp/call to allow offset from esp
I found an example of a generated jmp instruction that actually
referenced an address on the stack as an offset from esp:
  jmp DWORD PTR [esp+0x8]
This corner case had been in a TODO message for a long while in
the code, but I had been waiting for a real-world example before
implementing it.  The rewriter choked on this because I change esp
when I save the value of eax.  Since I use eax as a scratch
register, I need to push its old value, but in doing so I change
esp, and therefore any references to esp must be offset by the 4
bytes I changed esp with that push.

I wrote a test involving a few different encodings of jmp and call
instructions that use esp as a base.  I used the test number of
a test that has been broken for a while, and that violates some
assumptions I have about the address distribution of jit code.  I
don't plan to change miniverse to work with the broken test in the
near future, so I decided to move that test to a different name.