Fixed bug with cross-region jumps in initial jit regions
Since I wait until the first time the program attempts to run jit
code to rewrite it, I don't proactively rewrite any of the initial
jit regions prior to this first execution.  After that, I rewrite
all regions as soon as the jit compiler sets them as executable.
However, in this initial phase, it was possible for there to be
multiple non-rewritten regions, which would not be rewritten until
they were first executed.  This led to problems with inter-region
jumps, as a jump into a non-rewritten region leads to an attempted
lookup for code that has no lookup entries yet generated.  This led
to a crash.

I have created a test for this case and a fix.  The fix is to
rewrite all non-rewritten regions as soon as a single non-rewritten
region is encountered.  This should only happen at the start, but
I cautiously check for this every time in the segfault handler.
Perhaps I only should need to check the first time rewriting needs
to be done, and all times after that the code should have been
proactively rewritten, but keeping things this way provides some
support for RWX jit engines that don't behave well and don't follow
good W^X practice.