Terminate on segfaults caused by miniverse, lookup region size up
Now if a segfault takes place in the same region as miniverse's
gen_code function, I exit with an error that prints the segfault
address and offset from the start of that region.  This helped me
track down a crash caused by a segfault in the rewriter.  When
enough jit regions have been allocated that the amount of memory
I pre-allocate for the lookup entries is not enough, the rewriter
tries to write to non-allocated memory.  I have discussed a
(slower) lookup strategy with Dr. Hamlen that would allow for a
two-layer lookup, with a smaller fixed-location lookup table
that would provide the fixed offsets for large regions of memory.
That would allow me to allocate memory as-needed.  However, with
the current solution, I can't simply keep allocating more memory,
because as a program runs, memory might be allocated in the way
of where I might need to expand the memory.  This restriction is
required by the fixed offset I use currently.

I encountered this crash on a test case allocating a rather large
number of jit code regions, and it exceeded the amount I had
allocated.  I simply increased the amount I allocated, and now all
(non-wasm and non asm.js) firefox jit tests work.  This does not
mean that once I start testing in the browser again that it will
work, but the octane benchmarks do appear to use a lot of resources
and don't crash.  Therefore, there is hope that things might work
as a proof of concept without having to implement the more complex
lookup scheme.

I may want to consider writing a test that allocates a large amount
of jit code (at least enough to cause the crash with the previous
memory allocation amount).  This test could be changed to increase
the amount of memory allocated if I encounter the problem again
with even higher memory usage in the future.