Restored address masking code
Now the originally intended mask that aligns all indirect jumps
to jump to 16-byte chunk boundaries is functional again.  I could
not re-implement this until after I changed the lookup entries
to be 4 bytes, because when it was still possible for entries to
overlap, I used a fallback in the segfault handler for overlapping
entries.  For that to work, I needed an unmasked (and unmangled)
address to be retrieved by the handler, so I couldn't modify the
addresses.

I noticed after adding masking back in that the code caller
function in some of the earlier tests did not actually correctly
align the call instruction, nor put the correct tag byte after the
call.  I have changed these tests to function the same as the newer
tests I wrote, which use the miniverse_spatcher.py script to
ensure that all calls are correctly aligned and are followed by
the right tag byte.  This tag byte was also restored to its
original value (0x90, nop), instead of what I had been using
temporarily (0x9b, fwait).

It would probably be beneficial to review some of the code I
have previously commented out for cleanup, as well as to review
the accuracy of some of my comments containing the assembly code
explanations for each type of rewritten instruction.