Support for 3-byte calls in miniverse_spatcher, test with clang
It turns out that Firefox's build process uses clang, which
produces different instructions than gcc.  This ended up resulting
in my assembly patching script not correctly identifying call
instructions or the start of functions.  I added the ability to
build the tests with clang so I could fix this.

I discovered that my handling of certain types of call instructions
would not have worked correctly if gcc had generated them, as the
underlying assemblers used by both compilers can generate a 3-byte
instruction encoding with a 1-byte offset that I didn't anticipate.
If an offset from a register is in the range -128 to 127, then it
generates this shorter encoding.  I attempt to parse the offset in
my script as an integer.  If it succeeds, I check if it's in this
range and set the encoding size accordingly.  If I can't parse it
for any reason (such as if the offset is some symbol), then I
assume the assembler will choose the 6-byte encoding.  This could
produce an incorrectly aligned call, but this will only lead to
a problem if the call instruction calls into jit code.