Added performance profiling instrumentation
I now print how long various rewriting components take.  The
actual output is a little unintuitive, as it's split into seconds
and nanoseconds due to the timespec struct used by clock_gettime.
I don't calculate when the nanoseconds count exceeds a second, so
the seconds count by itself isn't sufficient to get a decent
approximation of time taken.  However, for debug purposes, this
has already proven to be useful, as it appears that reallocating
memory for rewritten code is currently contributing to the majority
of the overhead in the rewriter.  I originally implemented the
memory reallocation in a naive and inefficient way, so improving
that will hopefully provide an easy speed gain.  This, plus another
potential tweak to the code handling (mostly) unmodified
instructions, are the optimizations I plan to make and test next.