Improved new code memory allocation, don't modify PIC instructions
The way I re-allocated memory for the rewritten code was far too
conservative, leading to the memory repeatedly needing to be
re-allocated and copied to the new memory region.  This was fixed
by doubling the region size on reallocation (instead of adding
only one page at a time!) and starting the initial size at 16 times
the size of the original code region.  I checked, and when running
on Firefox's jit engine, the rewritten region tends to be 16 or 32
times the size of the original (likely because it can vary between
being slightly less and slightly more than 16 times larger).  To
avoid almost ever needing to re-allocate, I could set the starting
size to 32 times, but that would very likely waste memory.  Now
that the memory allocation portion takes up less than 10% of the
time taken when rewriting, I plan to focus on some other areas for
now, but I may come back to this later.

I also removed some unnecessary changes from being made to all
instructions that don't fall into a type of control flow
instruction.  These instructions should generally pass through
unmodified except for alignment, but I had some PIC-specific code
that I feel should have been conditionally removed with an #ifndef
back when I added #define PUSH_OLD_ADDRESSES, changing call
instructions to push old addresses on the stack and thereby
avoiding the need to use heuristics to identify PIC.  The one other
call to is_pic was only done if PUSH_OLD_ADDRESSES was not defined,
but somehow this one instance had slipped through.  Now I added a
preprocessor directive to avoid adding this if PUSH_OLD_ADDRESSES
is defined.  However, I'm considering whether I should just remove
all the code that runs if this is not defined along with the define
itself, because that code hasn't been tested in a long time, will
not work in Firefox, and is likely more fragile in general.